Make commands definition module public to clean up docs page
Also move response-relevant types outside the definitions module.
